support moderators: stop depending on `lgPlayer`
We need to make sure that moderators can issue any moderator-relevant command (like `vote` sliding out the voting panel).

Previously, some of the commands depended on information from the player record (e.g., `chapterId`, `goalRepositoryURL`). However, now all of the checking for that is (rightfully) done on the game server and this dependency is no longer in place.

The only thing remaining-in-use by game-cli from the `lgPlayer` attribute is the id (to ensure that the user is a player, which we can just as easily do with lgUser.roles). This commit removes any dependency that game-cli has on an `lgPlayer` being passed-in.